+++ libs/stegcommon/ByteBuffer.h
@@ -41,7 +41,7 @@ public:
     void write_uint32(uint32_t value);
 
     uint8_t read_uint8();
-    uint8_t read_uint16();
+    uint16_t read_uint16();
     uint32_t read_uint32() ;
 
     void reset_cursor();
